🎨 Schema refactor.
